Tidy up the /lib directory
