+++ spec/lib/sufia/user_stat_importer_spec.rb
@@ -1,5 +1,4 @@
 require 'spec_helper'
-require_relative '../../../lib/sufia/user_stat_importer'
 
 describe Sufia::UserStatImporter do
   before do
